Organize DX11 backend structure and shader sources.
Shorten DX11 type/file naming, move shader source classes into a dedicated Shaders subfolder and namespace, and update backend wiring/docs for the new layout.
